+++ pyxem/utils/__init__.py
@@ -20,7 +20,7 @@
 
 
 from pyxem.utils.ransac_ellipse_tools import determine_ellipse
-from pyxem.utils.calibration_utils import find_diffraction_calibration
+from pyxem.utils.calibration import find_diffraction_calibration
 from pyxem.utils.plotting import plot_template_over_pattern
 from pyxem.utils import expt_utils
